Algorithm1/requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0
tqdm>=4.65.0
numba

# Data processing
pyarrow>=14.0.1  # for parquet files
fastparquet>=2023.10.1
ta
# Machine Learning
lightgbm>=3.3.0
scikit-learn>=1.0.0
optuna>=3.0.0
mlxtend>=0.23.1  # for rule mining
torch>=1.13.0
pytorch-forecasting>=0.10.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.18.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
joblib>=1.2.0
